+++ handlers/uiHandlers.js
@@ -1,0 +1,20 @@
+const { addHeightMemory } = require('../socketUtils');
+const { SOCKET_EVENTS } = require('../constants');
+
+// --- UIハンドラーのセットアップ ---
+function setupUIHandlers(socket, io, heightMemory) {
+
+  socket.on(SOCKET_EVENTS.HEIGHT_CHANGE, (height) => {
+
+    // 高さメモリに追加
+    const heightArray = addHeightMemory(heightMemory, socket.id, height);
+
+    // 高さメモリを全クライアントにブロードキャスト(TODO: 同じルームにいる人にだけ送信するのか検討)
+    io.emit(SOCKET_EVENTS.HEIGHT_CHANGE, heightArray);
+
+  });
+}
+
+module.exports = {
+  setupUIHandlers
+};
